Move gui/TextEditor to control/tools
