+++ src/core/control/Control.cpp
@@ -35,13 +35,13 @@
 #include "control/settings/SettingsEnums.h"                      // for Button
 #include "control/settings/ViewModes.h"                          // for ViewM..
 #include "control/tools/EditSelection.h"                         // for Edit...
+#include "control/tools/TextEditor.h"                            // for Text...
 #include "control/xojfile/LoadHandler.h"                         // for Load...
 #include "control/zoom/ZoomControl.h"                            // for Zoom...
 #include "gui/MainWindow.h"                                      // for Main...
 #include "gui/PageView.h"                                        // for XojP...
 #include "gui/PdfFloatingToolbox.h"                              // for PdfF...
 #include "gui/SearchBar.h"                                       // for Sear...
-#include "gui/TextEditor.h"                                      // for Text...
 #include "gui/XournalView.h"                                     // for Xour...
 #include "gui/XournalppCursor.h"                                 // for Xour...
 #include "gui/dialog/AboutDialog.h"                              // for Abou...
@@ -1918,6 +1918,8 @@ void Control::selectDefaultTool() {
         selectTool(toolHandler->getToolType());
     }
 }
+
+void Control::setFontSelected(const XojFont& font) { this->getWindow()->setFontButtonFont(font); }
 
 void Control::toolChanged() {
     ToolType type = toolHandler->getToolType();
