+++ src/core/control/tools/TextEditor.h
