+++ src/core/control/tools/TextEditorWidget.h
